`ByteSwap` instance for `Word128`
This also exports the `ByteSwap` methods from `Data.Memory.Endian`. The
`Storable` instance for `Word128` depends on `getSystemEndianness`,
`Data.Memory.ExtendedWords` needs to  import of `Data.Memory.Endian`. If
we wanted to define the `ByteSwap` instance for `Word128` in
`Data.Memory.Endian` instead, we would end up with cyclic dependencies.
This seemed like the cleaner solution.